fix: Item options not showing & improve map update performance
- Fix item options loading from database (add_option_param now creates new options)

- Validate option IDs before adding to items

- Add head_avatar_manager for avatar system

- Move map update task to dedicated thread to prevent blocking tokio runtime

- Update NPC templates and data game structures

- Add debug logging for player blob serial

+++ src/entities/npc_template.rs
@@ -9,9 +9,9 @@ pub struct Model {
     pub id: i32,
     #[sea_orm(column_name = "NAME")]
     pub name: String,
-    pub head: i32,
-    pub body: i32,
-    pub leg: i32,
+    pub head: i16,
+    pub body: i16,
+    pub leg: i16,
     pub avatar: Option<i32>,
 }
